Use correct return type as specified in the problem statement

+++ swift-challenges/Triangular-treasure/Triangular-treasureTests/FactorialTests.swift
@@ -8,14 +8,14 @@
 import XCTest
 
 class Factorial {
-    func factorial(_ n: Int) -> Int {
+    func factorial(_ n: Int) -> UInt64 {
         //the factorial of `0` is defined as `1`
         if n == 0 {
             return 1
         } else {
             // The function calls itself with `n-1` until it reaches `0`, where the factorial of `0` is defined as `1`.
             //The results are then multiplied together as the stack unwinds.
-            return n * factorial(n - 1)
+            return UInt64(n) * factorial(n - 1)
         }
     }
 }
